+++ osrframework/mailfy.py
@@ -29,7 +29,7 @@ __author__ = "Felix Brezo, Yaiza Rubio "
 __copyright__ = "Copyright 2016, i3visio"
 __credits__ = ["Felix Brezo", "Yaiza Rubio"]
 __license__ = "GPLv3+"
-__version__ = "v3.0"
+__version__ = "v3.1"
 __maintainer__ = "Felix Brezo, Yaiza Rubio"
 __email__ = "contacto@i3visio.com"
 
@@ -270,8 +270,7 @@ This is free software, and you are welcome to redistribute it under certain cond
         print "You can find all the information collected in the following files:"
         for ext in args.extension:
             # Showing the output files
-            print "\t-" + fileHeader + "." + ext
-
+            print "\t-" + fileHeader + "." + ext"
     # Showing the execution time...
     if not args.quiet:
         print