+++ osrframework/wrappers/metacafe.py
@@ -3,7 +3,7 @@
 #
 ##################################################################################
 #
-#    Copyright 2015 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
+#    Copyright 2016 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
 #
 #    This program is part of OSRFramework. You can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -81,7 +81,7 @@ class Metacafe(Platform):
         # Strings that will imply that the query number is not appearing
         self.notFoundText = {}
         #self.notFoundText["phonefy"] = []
-        self.notFoundText["usufy"] = ["Metacafe - Online Video Entertainment - Free video clips for your enjoyment"]
+        self.notFoundText["usufy"] = ["<span>Channel is temporarily not available</span>"]
         #self.notFoundText["searchfy"] = []        
         
         #########################