+++ osrframework/wrappers/stumbleupon.py
@@ -3,7 +3,7 @@
 #
 ##################################################################################
 #
-#    Copyright 2015 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
+#    Copyright 2016 Félix Brezo and Yaiza Rubio (i3visio, contacto@i3visio.com)
 #
 #    This program is part of OSRFramework. You can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -81,7 +81,7 @@ class Stumbleupon(Platform):
         # Strings that will imply that the query number is not appearing
         self.notFoundText = {}
         #self.notFoundText["phonefy"] = []
-        self.notFoundText["usufy"] = ["<title>Explore more. Web pages, photos, and videos | StumbleUpon.com</title>"]
+        self.notFoundText["usufy"] = ["<h1><mark>Hmmm</mark>we can't find the page you're looking for.</mark></h1>"]
         #self.notFoundText["searchfy"] = []        
         
         #########################